\name{capitalize}
\alias{capitalize}
\title{Capitalize strings}
\description{Capitalizes characters that begin strings, or that follow a
  character that is not digit, underscore, or letter.
}
\usage{
capitalize(strings)
}
\arguments{
  \item{strings}{character vector}
}
\value{A character vector like \code{strings} but capitalized.
}
\author{ Jeff Hallman }
\seealso{ \code{\link{toupper}} and \code{\link{tolower}}}
\keyword{character}
